port: split msg-socket layer from server thread handling
